pyproject.toml
[tool.poetry]
name = "nextcord-test"
version = "0.1.0"
description = "A test bot to test bots with testing bots on nextcord"
authors = ["ooliver1 <oliverwilkes2006@icloud.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
nextcord = {extras = ["speed"], version = "^2.0.0-beta.2"}
uvloop = "^0.16.0"
ooliver-botbase = "^1.17.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
pre-commit = "^2.19.0"
taskipy = "^1.10.2"

[tool.taskipy.tasks]
precommit = { cmd = "pre-commit install --install-hooks" }
lint = { cmd = "pre-commit run --all-files"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
